MVP-5815-2: differ TargetListItem to 3 components: TargetListItemToggle/ TargetListItemForm
